Remove un-used code and fixes linting issues
Some un-used code was left over as part of the changes for v2.0.0
this change removes the remnants.

Also used [shellcheck](http://www.shellcheck.net/) to lint the code
and applied the code changes to resolve the identified issues.

One notable item was due to the use of a variable `split` which is an
actual bash command [split](https://ss64.com/bash/split.html). The
variable was renamed to `parts`.