Add lint check to Makefile and travis
Adds lint task to Makefile
Adds use of `make lint` to travis via docker service.

Resolves the last lint issue within test script.
Adds ignore only lint issue within semver. The lint issue is because
the variable is set using `eval` in another function which results
in shellcheck being unable to determine when/if the variable `parts`
was initiatlized with a value.