"release" subcommand fixed for "get", documentation updated/corrected
The "release" subcommand now works correctly with the "get" command.
It works as specified in the public API for semver 2.1.0 (README.md),
so this change does not impact the public API: it is a bug fix. An
example and test case has been added to the README.md to cover "get
release".

The public documentation for semver 2.1.0 does not specify a "bump
release" command. However, an example of this command is present in
the 2.1.0 README.md and "bump release" is specified in the script usage
message. Assuming this, the change to the public documentation in this
commit should not be considered as changing the public API (again, this
is a documentation bug fix, not a minor version increment).

README.md has been updated to correspond exactly to the script's usage
message. Not only is the release subcommand correctly documented, but
two spelling errors have been corrected and the regular expression has
been expanded.