semver tool compliant with semver.org spec version 2.0.0 (#37)
semver tool compliant with semver.org spec version 2.0.0

Change summary:

1. SEMVER_REGEX

SEMVER_REGEX now enforces the rule that numeric id's in the pre-release part
may not have leading zero's.  Alphanumeric id's may have leading zeros.
With the exception of the (permissive) acceptance of a leading "v" or "V",
the RE is now compliant.

The re-written SEMVER_REGEX has been decomposed into defined
sub-expressions to ease understanding of the code and to remove
redundancies.  However, the usage message and public API retain the fully
expanded RE which is now longer and more complex.

Because of changes in the RE groupings, the indexes to captured groups
in BASH_REMATCH have changed.

The SEMVER_REGEX is essentially the same as the Perl version published
by semver.org.  See: semver.org and https://regex101.com/r/vkijKf/1/

The only deltas are that:
1. The "\d" digit character class is not used.  "[0-9]" is used here
   instead.
2. Paren's used to capture sub-matches capture the leading "-" and "+"
   here due to implementation needs of the script. The semver.org regex
   has paren's to capture the pre-release and build parts without these
   leading characters.

The two RE's express exactly the same set of version strings.

2. Version Comparison

The comparison command has been completely reworked to comply with the
2.0.0 spec.  In particular, pre-release ids are compared one against the
other, the comparison is based on whether or not they are numeric or
not, numeric vs. non-numeric always has numeric at lower precedence, the
longer of otherwise identical pre-release parts has higher precedence.

The comparison implementation has been factored into functions.
The general style is the use of "truth tables" rather than
if-then-else statements. Efficiency is really not important in this
tool, but opportunities exist to boost/reorder/eliminate the lines
in the truth tables (e.g. boost the nat-2-nat compare) to shorten the
code and execution paths. Readability--more important here!--would
suffer.

The code to pass arrays as parameters to "compare-fields" is a bit
magic, but references to the technique can be found on the web.
See, e.g. https://stackoverflow.com/questions/11180714/how-to-iterate-over-an-array-using-indirect-reference

Lint complaints about non-quoted expansion in "compare-version" is
surpressed because we need word expansion to form the arrays of
pre-release fields.  The safety is insured due to the validation of the
contents by regular expression matching in "validate-version".

3. Documentation updates to semver 2.0.0 semantics

4. Version

The version has been set to 3.0.0-rc1 as this change to semver has
non-backward compatible changes:
1. Certain versions compare differently due to ident by ident comparison
   and the use of numeric vs ASCII sort order compares.
2. Certain versions are now rejected for not matching the semver regex:
   leading zeros are not allowed.

5. Testing

This commit has been tested against the "bats" unit tests and all tests
pass.  However, the commit does not depend on these tests being present.
The commit can be merged/rebased onto the unit test commit or
vice-versa.

Note that without this commit applied to the release branch (master),
the bats unit tests will fail (showing non-compliance with 2.0.0).
Applying this release makes the tests pass.