docs: replace "tag" badge with "release" badge (#56)
The release badge displays a label dynamically generated from the
latest github release.  Only stable (non pre-release) releases are
considered.  The URL/markdown reflects the recommended form according
to shields.io.

The "click-through" URL is likewise dynamic and points to the latest
release on github. That is, the label and URL are automatically kept
in sync. The previous "tag" badge used a static URL
that needed to be maintained to stay in sync with the dynamic label.

Note that "latest" is determined by date: what the github api
considers latest.

issues: #49