Add specific links to adsf and semver plugin.
The reader is directed to the adsf and semver plugin documentation
and README for more detailed examples and options.  The (previous)
single install line was too simple and could mis-lead the reader.
The reader is now referred to the example steps in the README and
the documentation covering version selection options (e.g. global
vs local, revert to system ..)