docs: Factor out package manager installation notes (issue #67)
The installation of "semver" via package/version/dependency
managers has been moved to the end of the README file.
The descriptions have been shortened and the style/substance
describing the two (known) managers is now consistent.

The details of how to set up the package managers has been
removed. Instead, the reader is referred to the respective web
sites. In this way, the semver-tool project does not duplicate
information and is not burdened with keeping it accurate,
complete, and up to date.

For example, the reader is directed to the asdf and semver
plugin documentation and README for detailed examples
and options: installation, version selection, etc.

This change addresses github issue #67 (and PR #65).

As well, the basic installation description has been
lightly edited.  And, the README file headline list now mentions
the recently added "validate" command.

Refs: #67, #65