Specify SPDX tag for semver
The semver utility is not widely packaged by distributions. As such
projects may wish to incorporate the script directly as a helper to
support packaging and versioning of a project.

In such cases, it is important to ensure that the licence of the semver
utility is maintained.

Provide an Apache-2.0 SPDX tag [0][1] within the semver script to ensure
the licencing is clear.

[0] https://spdx.dev/ids/
[1] https://spdx.org/licenses/Apache-2.0.html

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>