"prerelease" keyword supported in bump and get (issue #51) (#69)
The "diff" subcommand returns the keyword "prerelease" when it
is the pre-release part that differs. This keyword--different from
"prerel"--means using the output of "diff" is awkward to use as
input to "bump" or "get".  This change creates a synonym for "prerel",
namely "prerelease" in the bump and get subcommands, fixing the
output/input awkwardness without breaking backwards compatibility.

The implementation is simple: normalize the part keyword to a
canonical string ("prerel").  This is completely transparent to
the external API.

The README and usage string have been updated accordingly.

The README and usage string now list each subcommand on a separate
line for readability: the previous list of subcommand choices being
too long.

The order of the subcommands listed has been changed: the two
"complex" subcommands ("bump" and "get") appear ahead of the
others.

The changes pass all the existing unit tests: no regression or
backward compatibility issues. Two new test cases added to cover
use of this alternative keyword.

When released, these changes require bumping semver's minor version.
As a placeholder, the version string has been set to 3.4.0.