ci.yaml workflow drives "make" defined stable tests.
The GitHub action on pushes and pull requests has been
substantially simplified: on ubuntu-latest, check out the code
and run "make test-stable".

As noted in #74, the "asdf" install of a 'semver' release does not
test anything relevant to the push or PR.  It has been removed.

To unify testing between the CI and non-CI contexts, all tests are
driven by the same Makefile.  Hence, we don't have an "asdf" environment
set up in CI and something else in the local environment.  It is not
reasonable to impose "asdf" on developers, although this might be the
best tool for exploring new versions of tools.