README.md updated with development and test environment info.
The README file now explains more about the test environment (use and
design), the test tools, dependencies, and how one might set them up.

To motivate this section of the README--directed towards contributors
rather than consumers of semver--a new section was added covering
contributions to the project.  This includes guidelines and expectations
for contributors (sort of outside the scope of issue #74).

A typical development scenario is presented, running from working on
a specific change up through submitting a PR that will be subject to the
CI workflow.

A section was added suggesting that "asdf" might help in setting up
a local testing environment.