DRY and consistent testing framework, addresses issue #74 (#75)
* DRY dependency management and unified test/lint execution added.

A new file ("DEPENDENCIES") becomes the single location to specify
dependencies.  At present, these are only test tool dependencies.
The file format is that of shell variable assignments.  These can
be "included" in a Makefile to parameterize make's.  Project names
(a la GitHub), releases (i.e. versions), Docker image names, and
Docker tags are defined for each tool.

The "shellcheck" tool now has a well defined version: 0.8.0

The "bats" tool has been moved from 0.6.0 to 0.8.0

The "test/documentation-test" is now run as part of the complete test
suite. Strangly, it was previously omitted.

The Makefile has been expanded and fully parameterized.  The
fundamental design principle is to eliminate the difference between
local/development testing and automated CI (e.g. GitHub actions)
testing.  First, the arguments to be passed to test tools are defined
once in the Makefile and each is used in two places: to execute "local"
tests, and to execute "stable" tests.  Hence, the tests are the same,
but stable tests use well defined Docker images of the tool (specified
in DEPENDENCIES).  The local tests use whatever version of the test
tool is defined in the environment: a developer is free to work with
whatever tool version is appropriate and is responsible for providing
the tool (e.g. via "asdf", via manual installation, via a package
manager).

The names of the Makefile targets make this clear.

Note that with developer provided tools ("local"), these tools can be
used outside of the "make" tool.  As well, if Docker is installed,
configured, and running; the developer can execute the "stable" tests
locally (and/or use Docker "run" directly).  Hence, with Docker
installed, there is no reason to install the individual tools if so
desired.

While "make" is used mostly for testing (there is no building), an
"install" target is still available.

A user or developer downloading/cloning the semver-tool project and
naively typing "make" would most likely be confronted with errors (i.e.
tools or Docker not found).  As a courtesy to these users, the default
target is "semver" which informs the user that everything is already
built and points to the "help" target for further details.

Note that this Makefile design is completely independent of any CI
scripts. Of course, the desire is that this same test kit will be used
by CI/automated testing.  In particular, the CI pipeline will eventually
invoke "make test-stable".

* ci.yaml workflow drives "make" defined stable tests.

The GitHub action on pushes and pull requests has been
substantially simplified: on ubuntu-latest, check out the code
and run "make test-stable".

As noted in #74, the "asdf" install of a 'semver' release does not
test anything relevant to the push or PR.  It has been removed.

To unify testing between the CI and non-CI contexts, all tests are
driven by the same Makefile.  Hence, we don't have an "asdf" environment
set up in CI and something else in the local environment.  It is not
reasonable to impose "asdf" on developers, although this might be the
best tool for exploring new versions of tools.

* README.md updated with development and test environment info.

The README file now explains more about the test environment (use and
design), the test tools, dependencies, and how one might set them up.

To motivate this section of the README--directed towards contributors
rather than consumers of semver--a new section was added covering
contributions to the project.  This includes guidelines and expectations
for contributors (sort of outside the scope of issue #74).

A typical development scenario is presented, running from working on
a specific change up through submitting a PR that will be subject to the
CI workflow.

A section was added suggesting that "asdf" might help in setting up
a local testing environment.

* Removed trigger on pushes to issue_74 branch.

The workflow is cleaned up (after being used for testing) so that it
only triggers on PR's and pushes to 'master'